+++ functions/AllowEdit.fnc.php
@@ -49,7 +49,8 @@ function AllowEdit( $modname = false )
 
 	// Student / User Info tabs.
 	if ( ( $modname === 'Students/Student.php'
-		|| $modname === 'Users/User.php' )
+			|| $modname === 'Users/User.php' )
+		&& ( isset( $_REQUEST['modname'] ) && $modname === $_REQUEST['modname'] )
 		&& isset( $_REQUEST['category_id'] ) )
 	{
 		$modname = $modname . '&category_id=' . $_REQUEST['category_id'];
@@ -94,7 +95,7 @@ function AllowUse( $modname = false )
 	// Student / User Info tabs.
 	if ( ( $modname === 'Students/Student.php'
 			|| $modname === 'Users/User.php' )
-		&& $modname === $_REQUEST['modname']
+		&& ( isset( $_REQUEST['modname'] ) && $modname === $_REQUEST['modname'] )
 		&& isset( $_REQUEST['category_id'] ) )
 	{
 		$modname = $modname . '&category_id=' . $_REQUEST['category_id'];